backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
vllm>=0.1.4
loguru>=0.7.0
sse-starlette>=1.6.0
python-dotenv>=1.0.0
httpx>=0.24.0
tiktoken>=0.4.0
starlette>=0.27.0
sentencepiece>=0.1.99
protobuf>=4.23.0
einops>=0.6.1
redis>=4.5.5
pynvml>=11.5.0
